REoptLite.jl

Renewable Energy Optimization and Integration

Note

This package is currently under development and not necessarily stable. It contains a subset of the REopt Lite API capabilities.

REoptLite.build_reopt! — Method

build_reopt!(m::JuMP.AbstractModel, fp::String)

Add variables and constraints for REopt model. fp is used to load in JSON file to construct REoptInputs.

REoptLite.Scenario — Method

Scenario(d::Dict)

Constructor for Scenario struct, where d has upper-case keys:

Site (required)

ElectricTariff (required)

ElectricLoad (required)

PV (optional, can be Array)

Storage (optional)

ElectricUtility (optional)

Financial (optional)

Generator (optional)

All values of d are expected to Dicts except for PV, which can be either a Dict or Dict[].

struct Scenario
    site::Site
    pvs::Array{PV, 1}
    storage::Storage
    electric_tariff::ElectricTariff
    electric_load::ElectricLoad
    electric_utility::ElectricUtility
    financial::Financial
    generator::Generator
end