REopt Inputs

Inputs to run_reopt can be provided in one of three formats:

a file path (string) to a JSON file,

a Dict, or

using the REoptInputs struct

The first option is perhaps the most straightforward one. For example, the minimum requirements for a JSON scenario file would look like:

{
    "Site": {
        "longitude": -118.1164613,
        "latitude": 34.5794343
    },
    "ElectricLoad": {
        "doe_reference_name": "MidriseApartment",
        "annual_kwh": 1000000.0,
        "city": "Boulder"
    },
    "ElectricTariff": {
        "urdb_label": "5ed6c1a15457a3367add15ae"
    }
}

The order of the keys do not matter. Note that this scenario does not include any energy generation technologies and therefore the results can be used as a baseline for comparison to scenarios that result in cost-optimal generation technologies.

To add PV to the analysis simply add a PV key with an empty dictionary (to use default values):

{
    "Site": {
        "longitude": -118.1164613,
        "latitude": 34.5794343
    },
    "ElectricLoad": {
        "doe_reference_name": "MidriseApartment",
        "annual_kwh": 1000000.0,
        "city": "Boulder"
    },
    "ElectricTariff": {
        "urdb_label": "5ed6c1a15457a3367add15ae"
    },
    "PV": {}
}

This scenario will consider the option to purchase a solar PV system to reduce energy costs, and if solar PV can reduce the energy costs then REopt Lite will provide the optimal PV capacity (assuming perfect foresight!). To use other than default values for PV see the PV struct definition. For example, the site under consideration might have some existing PV capacity to account for, which can be done by setting the existing_kw key to the appropriate value.

Scenario

The Scenario struct captures all of the objects that can be included in a scenario.json:

REoptLite.Scenario — Type

Scenario(d::Dict)

Constructor for Scenario struct, where d has upper-case keys:

Site (required)

ElectricTariff (required)

ElectricLoad (required)

PV (optional, can be Array)

Storage (optional)

ElectricUtility (optional)

Financial (optional)

Generator (optional)

All values of d are expected to Dicts except for PV, which can be either a Dict or Dict[].

struct Scenario
    site::Site
    pvs::Array{PV, 1}
    storage::Storage
    electric_tariff::ElectricTariff
    electric_load::ElectricLoad
    electric_utility::ElectricUtility
    financial::Financial
    generator::Generator
end

REoptInputs

REoptLite.REoptInputs — Type

REoptInputs(fp::String)

Use fp to load in JSON scenario:

function REoptInputs(fp::String)
    s = Scenario(JSON.parsefile(fp))
    REoptInputs(s)
end

Useful if you want to manually modify REoptInputs before solving the model.

ElectricLoad

REoptLite.ElectricLoad — Type

ElectricLoad(;
    loads_kw::Union{Missing, Array{<:Real,1}} = missing,
    year::Int = 2020,
    doe_reference_name::Union{Missing, String} = missing,
    city::Union{Missing, String} = missing,
    annual_kwh::Union{Real, Nothing} = nothing,
    monthly_totals_kwh::Array{<:Real,1} = Real[],
    critical_loads_kw::Union{Missing, Array{Real,1}} = missing,
    loads_kw_is_net::Bool = true,
    critical_loads_kw_is_net::Bool = false,
    critical_load_pct::Real = 0.5
)

Must provide either loads_kw or [doe_reference_name and city] or doe_reference_name.

When only doe_reference_name is provided the Site.latitude and Site.longitude are used to look up the ASHRAE climate zone, which determines the appropriate DoE Commercial Reference Building profile.

When using the [doe_reference_name and city] option, choose city from one of the cities used to represent the ASHRAE climate zones:

Albuquerque

Atlanta

Baltimore

Boulder

Chicago

Duluth

Fairbanks

Helena

Houston

LosAngeles

Miami

Minneapolis

Phoenix

SanFrancisco

Seattle

and doe_reference_name from:

FastFoodRest

FullServiceRest

Hospital

LargeHotel

LargeOffice

MediumOffice

MidriseApartment

Outpatient

PrimarySchool

RetailStore

SecondarySchool

SmallHotel

SmallOffice

StripMall

Supermarket

Warehouse

FlatLoad

Each city and doe_reference_name combination has a default annual_kwh, or you can provide your own annual_kwh or monthly_totals_kwh and the reference profile will be scaled appropriately.

Financial values

REoptLite.Financial — Type

Financial

Base.@kwdef struct Financial
    om_cost_escalation_pct::Float64 = 0.025
    elec_cost_escalation_pct::Float64 = 0.023
    offtaker_tax_pct::Float64 = 0.26
    offtaker_discount_pct = 0.083
    two_party_ownership::Bool = false
    owner_tax_pct::Float64 = 0.26
    owner_discount_pct::Float64 = 0.083
    analysis_years::Int = 25
    macrs_five_year::Array{Float64,1} = [0.2, 0.32, 0.192, 0.1152, 0.1152, 0.0576]  # IRS pub 946
    macrs_seven_year::Array{Float64,1} = [0.1429, 0.2449, 0.1749, 0.1249, 0.0893, 0.0892, 0.0893, 0.0446]
    VoLL::Union{Array{R,1}, R} where R<:Real = 1.00
    microgrid_premium_pct::Float64 = 0.3
end

ElectricUtility

REoptLite.ElectricUtility — Type

ElectricUtility

Base.@kwdef struct ElectricUtility
    outage_start_timestep::Int=0  # for modeling a single outage, with critical load spliced into the baseline load ...
    outage_end_timestep::Int=0  # ... utiltity production_factor = 0 during the outage
    allow_simultaneous_export_import::Bool=true  # if true the site has two meters (in effect)
    # variables below used for minimax the expected outage cost,
    # with max taken over outage start time, expectation taken over outage duration
    outage_start_timesteps::Array{Int,1}=Int[]  # we minimize the maximum outage cost over outage start times
    outage_durations::Array{Int,1}=Int[]  # one-to-one with outage_probabilities, outage_durations can be a random variable
    outage_probabilities::Array{Real,1}=[1.0]
    outage_timesteps::Union{Missing, UnitRange} = isempty(outage_durations) ? missing : 1:maximum(outage_durations)
    scenarios::Union{Missing, UnitRange} = isempty(outage_durations) ? missing : 1:length(outage_durations)
end

Generator

REoptLite.Generator — Type

Generator

struct with inner constructor:

function Generator(;
    existing_kw::Real=0,
    min_kw::Real=0,
    max_kw::Real=1.0e6,
    cost_per_kw::Real=500.0,
    om_cost_per_kw::Real=10.0,
    om_cost_per_kwh::Float64=0.0,
    fuel_cost_per_gallon::Float64 = 3.0,
    fuel_slope_gal_per_kwh::Float64 = 0.076,
    fuel_intercept_gal_per_hr::Float64 = 0.0,
    fuel_avail_gal::Float64 = 660.0,
    min_turn_down_pct::Float64 = 0.0,
    only_runs_during_grid_outage::Bool = true,
    sells_energy_back_to_grid::Bool = false
)

ElectricTariff

REoptLite.ElectricTariff — Type

ElectricTariff

function ElectricTariff(;
    urdb_label::String="",
    urdb_response::Dict=Dict(),
    year::Int=2020,
    time_steps_per_hour::Int=1,
    NEM::Bool=false,
    wholesale_rate::T=nothing, 
    monthly_energy_rates::Array=[],
    monthly_demand_rates::Array=[],
) where {T <: Union{Nothing, Int, Float64, Array}, S <: Union{Nothing, Int, Float64, Array}}

Settings

REoptLite.Settings — Type

Settings

Data struct for top-level Scenario settings. Captures inputs that do not logically fall under any of the other data structs.

Base.@kwdef struct Settings
    time_steps_per_hour::Int = 1
end