Methods

REoptLite.Scenario — Method

Scenario(d::Dict)

Constructor for Scenario struct, where d has upper-case keys:

Site (required)

ElectricTariff (required)

ElectricLoad (required)

PV (optional, can be Array)

Wind (optional)

Storage (optional)

ElectricUtility (optional)

Financial (optional)

Generator (optional)

All values of d are expected to be Dicts except for PV, which can be either a Dict or Dict[].

struct Scenario
    settings::Settings
    site::Site
    pvs::Array{PV, 1}
    wind::Wind
    storage::Storage
    electric_tariff::ElectricTariff
    electric_load::ElectricLoad
    electric_utility::ElectricUtility
    financial::Financial
    generator::Generator
end

run_reopt

REoptLite.run_reopt — Function

run_reopt(m::JuMP.AbstractModel, fp::String)

Solve the model using the Scenario defined in JSON file stored at the file path fp.

run_reopt(m::JuMP.AbstractModel, d::Dict)

Solve the model using the Scenario defined in dict d.

run_reopt(m::JuMP.AbstractModel, s::AbstractScenario)

Solve the model using a Scenario or BAUScenario.

run_reopt(t::Tuple{JuMP.AbstractModel, AbstractScenario})

Method for use with Threads when running BAU in parallel with optimal scenario.

run_reopt(ms::AbstractArray{T, 1}, fp::String) where T <: JuMP.AbstractModel

Solve the Scenario and BAUScenario in parallel using the first two (empty) models in ms and inputs defined in the JSON file at the filepath fp.

run_reopt(ms::AbstractArray{T, 1}, d::Dict) where T <: JuMP.AbstractModel

Solve the Scenario and BAUScenario in parallel using the first two (empty) models in ms and inputs from d.

run_reopt(ms::AbstractArray{T, 1}, p::REoptInputs) where T <: JuMP.AbstractModel

Solve the Scenario and BAUScenario in parallel using the first two (empty) models in ms and inputs from p.

build_reopt!

REoptLite.build_reopt! — Method

build_reopt!(m::JuMP.AbstractModel, fp::String)

Add variables and constraints for REopt model. fp is used to load in JSON file to construct REoptInputs.